requirements.txt
annotated-types==0.5.0
anyio==3.6.2
click==8.1.3
et-xmlfile==1.1.0
fastapi==0.100.0
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.6
pydantic[email]
pydantic_core==2.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
passlib[bcrypt]
python-jose==3.3.0
python-multipart==0.0.6
pillow==10.0.0
reportlab==4.0.4
PyPDF2==3.0.1